feat: http functions refactoring
