+++ pyeudiw/federation/exceptions.py
@@ -1,6 +1,4 @@
-class HttpError(Exception):
-    pass
-
+from pyeudiw.tools.exceptions import HttpError
 
 class TrustChainHttpError(HttpError):
     pass
